Changed Batch to Subset as it pertains to worksheet. Added sort_order field to test_worksheet_item. Added qc_system_user_id and qc_started_date to worksheet_analysis. Added instrument to worksheet. Added code to change analysis status from requeue to initiated when it is added to a new worksheet. Added fetchActiveByName to Instrument. Added overrides for QCs on worksheet. Added code to not transfer values from the worksheet to analysis when the analysis status is set to requeue. Changed the worksheet analysis table on the worksheet tab of the test screen to an exclusion list. Added code to modify instrument log entries when an instrument is associated with a worksheet. Removed collected date requirement from sample validation.

+++ ejb/src/org/openelis/bean/WorksheetCreationBean.java
@@ -156,7 +156,7 @@ public class WorksheetCreationBean implements WorksheetCreationRemote {
     /*
      * Compute the number of days before the analysis is expected to be finshed
      */
-    private Long computeDueDays(Datetime received, Integer expectedDays) {
+    private Integer computeDueDays(Datetime received, Integer expectedDays) {
         long     due;
         Datetime now, expectedDate;
         
@@ -172,7 +172,7 @@ public class WorksheetCreationBean implements WorksheetCreationRemote {
         // convert from milliseconds to days
         due = due / 1000 / 60 / 60 / 24;
         
-        return due;
+        return (int)due;
     }
     
     /*
